refactor: DiffUtil 위치 수정
