+++ app/src/main/java/com/mashup/ipdam/entity/history/History.kt
@@ -1,5 +1,6 @@
 package com.mashup.ipdam.entity.history
 
+import androidx.recyclerview.widget.DiffUtil
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
@@ -18,6 +19,18 @@ data class History(
 
     override fun hashCode(): Int {
         return address.hashCode()
+    }
+
+    companion object {
+        val DiffCallback = object : DiffUtil.ItemCallback<History>() {
+            override fun areItemsTheSame(oldItem: History, newItem: History): Boolean {
+                return oldItem.id == newItem.id
+            }
+
+            override fun areContentsTheSame(oldItem: History, newItem: History): Boolean {
+                return oldItem == newItem
+            }
+        }
     }
 }
