+++ app/src/main/java/com/mashup/ipdam/ui/search/adapter/history/HistoryAdapter.kt
@@ -2,7 +2,6 @@ package com.mashup.ipdam.ui.search.adapter.history
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.mashup.ipdam.databinding.ItemHistoryBinding
 import com.mashup.ipdam.ui.search.SearchViewModel
@@ -10,7 +9,7 @@ import com.mashup.ipdam.entity.history.History
 
 class HistoryAdapter(
     private val searchViewModel: SearchViewModel
-) : ListAdapter<History, HistoryViewHolder>(HistoryDiffCallback) {
+) : ListAdapter<History, HistoryViewHolder>(History.DiffCallback) {
 
     init {
         setHasStableIds(true)
@@ -27,15 +26,4 @@ class HistoryAdapter(
     override fun onBindViewHolder(holder: HistoryViewHolder, position: Int) {
         holder.bind(getItem(position))
     }
-}
-
-object HistoryDiffCallback : DiffUtil.ItemCallback<History>() {
-    override fun areItemsTheSame(oldItem: History, newItem: History): Boolean {
-        return oldItem.id == newItem.id
-    }
-
-    override fun areContentsTheSame(oldItem: History, newItem: History): Boolean {
-        return oldItem == newItem
-    }
-
 }
