+++ app/src/main/java/com/mashup/ipdam/ui/search/adapter/kakao/PlaceAdapter.kt
@@ -2,14 +2,13 @@ package com.mashup.ipdam.ui.search.adapter.kakao
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.mashup.ipdam.databinding.ItemPlaceBinding
 import com.mashup.ipdam.entity.kakao.Places
 
 class PlaceAdapter(
     private val placeClickListener: PlaceClickListener
-) : ListAdapter<Places, PlaceViewHolder>(PlacesDiffCallback()) {
+) : ListAdapter<Places, PlaceViewHolder>(Places.DiffCallback) {
 
     init {
         setHasStableIds(true)
@@ -29,15 +28,5 @@ class PlaceAdapter(
 
     interface PlaceClickListener {
         fun onPlaceClick(position: Int)
-    }
-}
-
-class PlacesDiffCallback : DiffUtil.ItemCallback<Places>() {
-    override fun areItemsTheSame(oldItem: Places, newItem: Places): Boolean {
-        return oldItem.id == newItem.id
-    }
-
-    override fun areContentsTheSame(oldItem: Places, newItem: Places): Boolean {
-        return oldItem.id == newItem.id
     }
 }
